backend/requirements.txt
openai>=1.0.0
fastapi
uvicorn
requests
pydantic
PyYAML
diff-match-patch
